+++ composeUi/src/commonMain/kotlin/com/darkrockstudios/apps/hammer/common/projectroot/ProjectRootUi.kt
@@ -4,8 +4,10 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -25,6 +27,8 @@ import com.darkrockstudios.apps.hammer.common.compose.SetScreenCharacteristics
 import com.darkrockstudios.apps.hammer.common.compose.Ui
 import com.darkrockstudios.apps.hammer.common.compose.designsystem.HdBottomBarDestination
 import com.darkrockstudios.apps.hammer.common.compose.designsystem.HdNavRailDestination
+import com.darkrockstudios.apps.hammer.common.compose.ProjectShortcutHost
+import com.darkrockstudios.apps.hammer.common.compose.rememberStrRes
 import com.darkrockstudios.apps.hammer.common.compose.resources.get
 import com.darkrockstudios.apps.hammer.common.encyclopedia.BrowseEntriesFab
 import com.darkrockstudios.apps.hammer.common.encyclopedia.EncyclopediaUi
@@ -44,6 +48,8 @@ import com.darkrockstudios.apps.hammer.ic_encyclopedia
 import com.darkrockstudios.apps.hammer.ic_home
 import com.darkrockstudios.apps.hammer.ic_notes
 import com.darkrockstudios.apps.hammer.ic_timeline
+import com.darkrockstudios.apps.hammer.save_all_toast
+import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.vectorResource
 
 private val WIDE_SCREEN_THRESHOLD = 650.dp
@@ -83,7 +89,26 @@ fun ProjectRootUi(
 	rootSnackbar: RootSnackbarHostState,
 	navWidth: Dp = Dp.Unspecified,
 	modifier: Modifier = Modifier,
+	shortcutHost: ProjectShortcutHost? = null,
 ) {
+	val scope = rememberCoroutineScope()
+	val strRes = rememberStrRes()
+
+	if (shortcutHost != null) {
+		DisposableEffect(shortcutHost, component, rootSnackbar, scope, strRes) {
+			shortcutHost.bind(
+				startSync = { component.startProjectSync() },
+				saveAll = {
+					scope.launch {
+						component.storeDirtyBuffers()
+						rootSnackbar.showSnackbar(strRes.get(Res.string.save_all_toast))
+					}
+				},
+			)
+			onDispose { shortcutHost.unbind() }
+		}
+	}
+
 	SetScreenCharacteristics(WIDE_SCREEN_THRESHOLD) {
 		FeatureContent(
 			modifier.fillMaxSize(),
